# internal/modules/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 九小二 Backend API
  version: 1.0.0-cp1
  description: 九小二产品一期 Go 后端唯一公开契约，覆盖交易、履约、打印、通知、账号运营与酒类合规。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /admin/dispatch/policies:
    get:
      summary: 查询混合派单策略版本
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageToken"
        - { name: filter, in: query, schema: { type: string, maxLength: 512 } }
        - { name: order_by, in: query, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/Page" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      summary: 创建派单策略草稿
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DispatchPolicyCreateReq" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /admin/dispatch/policies/{id}/validate:
    post:
      summary: 校验派单策略
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody: { $ref: "#/components/requestBodies/ExpectedVersion" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /admin/dispatch/policies/{id}/publish:
    post:
      summary: 发布派单策略
      parameters:
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody: { $ref: "#/components/requestBodies/ExpectedVersion" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /admin/dispatch/policies/{id}/retire:
    post:
      summary: 停用派单策略
      parameters:
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody: { $ref: "#/components/requestBodies/ExpectedVersion" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /admin/dispatch/jobs:
    get:
      summary: 查询派单任务
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageToken"
        - { name: filter, in: query, schema: { type: string, maxLength: 512 } }
        - { name: order_by, in: query, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/Page" }
        default: { $ref: "#/components/responses/Problem" }
  /admin/dispatch/jobs/{id}:
    get:
      summary: 查询派单任务时间线
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
  /admin/dispatch/jobs/{id}/retry:
    post:
      summary: 重试需要人工处理的派单任务
      parameters:
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_version, reason_code, reason]
              properties:
                expected_version: { type: integer, minimum: 1 }
                reason_code: { type: string, maxLength: 64 }
                reason: { type: string, minLength: 2, maxLength: 255 }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    applicationBearer:
      type: http
      scheme: bearer
      bearerFormat: application_access JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
    PageToken:
      name: page_token
      in: query
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
...
  responses:
    Success:
      description: 请求成功
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessEnvelope" }
    Page:
      description: 分页请求成功
      content:
        application/json:
          schema: { $ref: "#/components/schemas/PageEnvelope" }
...
    Problem:
      description: 标准业务或系统错误
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...
  requestBodies:
    ExpectedVersion:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [expected_version]
            properties:
              expected_version: { type: integer, minimum: 1 }
...
  schemas:
...
    DispatchPolicyCreateReq:
      type: object
      required:
        - policy_code
        - scope_type
        - scope_id
        - mode
        - offer_ttl_seconds
        - grab_ttl_seconds
        - candidate_limit
        - offer_candidate_limit
        - heartbeat_fresh_seconds
        - location_fresh_seconds
        - max_location_accuracy_m
        - max_pickup_distance_m
        - max_active_orders_default
        - idle_full_score_seconds
        - score_weights
      properties:
        policy_code: { type: string, maxLength: 64 }
        scope_type: { type: string, enum: [global, city, shop] }
        scope_id: { type: string, maxLength: 64 }
        mode: { type: string, enum: [hybrid, auto, grab, manual] }
        auto_rounds: { type: integer, minimum: 0, maximum: 10 }
        offer_ttl_seconds: { type: integer, minimum: 5, maximum: 60 }
        grab_ttl_seconds: { type: integer, minimum: 5, maximum: 300 }
        candidate_limit: { type: integer, minimum: 1, maximum: 500 }
        offer_candidate_limit: { type: integer, minimum: 1, maximum: 20 }
        heartbeat_fresh_seconds: { type: integer, minimum: 15, maximum: 300 }
        location_fresh_seconds: { type: integer, minimum: 30, maximum: 600 }
        max_location_accuracy_m: { type: integer, minimum: 20, maximum: 1000 }
        max_pickup_distance_m: { type: integer, minimum: 500, maximum: 50000 }
        max_active_orders_default: { type: integer, minimum: 1, maximum: 20 }
        idle_full_score_seconds: { type: integer, minimum: 60, maximum: 86400 }
        rejection_cooldown_seconds: { type: integer, minimum: 0, maximum: 3600 }
        score_weights:
          type: object
          required: [distance, load, idle, freshness]
          properties:
            distance: { type: number, minimum: 0, maximum: 1 }
            load: { type: number, minimum: 0, maximum: 1 }
            idle: { type: number, minimum: 0, maximum: 1 }
            freshness: { type: number, minimum: 0, maximum: 1 }
    SuccessEnvelope:
      type: object
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data: { type: object, additionalProperties: true, nullable: true }
        request_id: { type: string }
...
    PageEnvelope:
      type: object
      required: [code, message, data, request_id]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data:
          type: object
          required: [items, next_page_token]
          properties:
            items:
              type: array
              items: { type: object, additionalProperties: true }
            next_page_token: { type: string }
        request_id: { type: string }
...
    ProblemDetails:
      type: object
      required: [type, title, status, error_code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        error_code:
          type: string
          description: 幂等写请求同 key 异参固定返回 IDEMPOTENCY_KEY_REUSED；同参请求仍处于有效处理租约时固定返回 IDEMPOTENCY_IN_PROGRESS。
        request_id: { type: string }